Fixed a bug which caused buffer overflow when encrypting files with a size of over 2 GBs.

+++ src/file.h
@@ -3,7 +3,7 @@
 
 #include "crypto.h"
 
-DWORD get_file_size(LPCSTR f_name);
+LONGLONG get_file_size(LPCSTR f_name);
 BOOL get_and_strip_iv(LPCSTR f_name, AES_KEY *a);
 BOOL get_and_not_strip_iv(LPCSTR f_name, AES_KEY *a);
 BOOL strip_file(LPCSTR f_name, int origin, long offset);
